+++ EstateAgency.Application/DTOs/ClientDto.cs
@@ -15,42 +15,15 @@ public class ClientDto
     /// <summary>
     /// Полное имя клиента в формате "Фамилия Имя Отчество"
     /// </summary>
-    public string FullName { get; set; } = string.Empty;
+    public required string FullName { get; set; }
 
     /// <summary>
     /// Номер паспорта клиента в формате "XXXX XXXXXX"
     /// </summary>
-    public string PassportNumber { get; set; } = string.Empty;
+    public required string PassportNumber { get; set; }
 
     /// <summary>
     /// Контактный телефон клиента в международном формате
     /// </summary>
-    public string PhoneNumber { get; set; } = string.Empty;
-}
-
-/// <summary>
-/// Содержит валидационные правила для входящих данных
-/// </summary>
-public class CreateClientDto
-{
-    /// <summary>
-    /// Полное имя клиента. Обязательное поле, максимум 100 символов
-    /// </summary>
-    [Required(ErrorMessage = "ФИО клиента обязательно")]
-    [StringLength(100, ErrorMessage = "ФИО не может превышать 100 символов")]
-    public string FullName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Номер паспорта в формате "XXXX XXXXXX"
-    /// </summary>
-    [Required(ErrorMessage = "Номер паспорта обязателен")]
-    [RegularExpression(@"^\d{4} \d{6}$", ErrorMessage = "Номер паспорта должен быть в формате 'XXXX XXXXXX'")]
-    public string PassportNumber { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Контактный телефон для связи с клиентом
-    /// </summary>
-    [Required(ErrorMessage = "Контактный телефон обязателен")]
-    [Phone(ErrorMessage = "Некорректный формат телефона")]
-    public string PhoneNumber { get; set; } = string.Empty;
+    public required string PhoneNumber { get; set; }
 }
