+++ EstateAgency.Application/DTOs/RequestDto.cs
@@ -21,7 +21,7 @@ public class RequestDto
     /// <summary>
     /// Полное имя клиента
     /// </summary>
-    public string ClientFullName { get; set; } = string.Empty;
+    public required string ClientFullName { get; set; }
 
     /// <summary>
     /// Идентификатор объекта недвижимости 
@@ -31,7 +31,7 @@ public class RequestDto
     /// <summary>
     /// Адрес объекта недвижимости
     /// </summary>
-    public string PropertyAddress { get; set; } = string.Empty;
+    public required string PropertyAddress { get; set; } 
 
     /// <summary>
     /// Тип заявки  
@@ -47,35 +47,4 @@ public class RequestDto
     /// Дата и время создания заявки 
     /// </summary>
     public DateTime CreatedDate { get; set; }
-}
-
-/// <summary> 
-/// Содержит валидационные правила для финансовых данных и связей 
-/// </summary>
-public class CreateRequestDto
-{
-    /// <summary>
-    /// Идентификатор клиента, создающего заявку 
-    /// </summary>
-    [Required(ErrorMessage = "Идентификатор клиента обязателен")]
-    public int ClientId { get; set; }
-
-    /// <summary>
-    /// Идентификатор объекта недвижимости, к которому относится заявка 
-    /// </summary>
-    [Required(ErrorMessage = "Идентификатор недвижимости обязателен")]
-    public int PropertyId { get; set; }
-
-    /// <summary>
-    /// Тип заявки 
-    /// </summary>
-    [Required(ErrorMessage = "Тип заявки обязателен")]
-    public RequestType Type { get; set; }
-
-    /// <summary>
-    /// Денежная сумма заявки
-    /// </summary>
-    [Required(ErrorMessage = "Сумма заявки обязательна")]
-    [Range(0.01, double.MaxValue, ErrorMessage = "Сумма должна быть больше 0")]
-    public decimal Amount { get; set; }
 }
