+++ EstateAgency.Application/Mappings/AutoMapperProfile.cs
@@ -21,8 +21,8 @@ public class AutoMapperProfile : Profile
         CreateMap<CreatePropertyDto, Property>();
 
         CreateMap<Request, RequestDto>()
-            .ForMember(dest => dest.ClientFullName, opt => opt.MapFrom(src => src.Client.FullName))
-            .ForMember(dest => dest.PropertyAddress, opt => opt.MapFrom(src => src.Property.Address));
+            .ForMember(dest => dest.ClientFullName, opt => opt.MapFrom(src => src.Client != null ? src.Client.FullName : null))
+            .ForMember(dest => dest.PropertyAddress, opt => opt.MapFrom(src => src.Property != null ? src.Property.Address : null));
         CreateMap<CreateRequestDto, Request>();
     }
 }
